Create README.md
Application Description:

This console application was developed as a project for the course "Object-Oriented Programming – Internet Technologies". The application allows users to manage a contact list with the following features:
•	View all contacts: Displays all the contacts in the address book.
•	Add a new contact: Allows the user to add a new contact by providing necessary details such as name, phone number, address, email, and date of birth.
•	Search for a contact: Enables the user to search for a contact by name and view their details.
•	Edit a contact: Allows the user to edit the details of an existing contact by searching for the contact by name.
•	Delete a contact: Enables the user to delete a contact from the address book by searching for the contact by name.
•	Exit the application: Saves the changes and exits the application.
Serialization
The application uses Java's serialization mechanism to save and load contact data to and from a file named AddressContact.ser. This ensures that all contact information is preserved between sessions.
Author
•	Michalis Papanikolas
o	Year: 2021